pyproject.toml
[project]
name = "week-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.1",
]
